Setup repository
After cloning, the repository is setup for this particular project. Unnecessary files/folders are removed, relevant simulations codes are placed in the runs directory, the README.md file was edited to better reflect the content of this project, miscellaneous auxiliary scripts are inserted in relevant places, and, most importantly, certain macros are set so the code is optimized for the specific application in mind.