updated radially varying diffusion class
In the diffusion class *DiffusionKineticEnergyRadialDistancePowerLaw* introduced a *down_stream_index* which will take care of a different variation in diffusion downstream. When the index is "0" diffusion coefficients vary radially and dependent on kinetic energy based upon the power_law_Index of "r" and kinetic energy. But when this index is 1, Diffusion coefficient vary differently for r < r_dn and r> r_dn where *r_dn* is the starting distance of the downstream region.