Reorganized file structure
The general file structure was reorganized in preparation for a spherical shock code. Note that, at the moment, the codes in the runs_spherical_shock directory are just copies of those in runs_planar_shock, so no actual spherical shock code is currently in this repository. The README.md and configure.sh files were updated accordingly and the former was expanded for greater clarity. Minor changes were made to the use of the time array index argument in the backward planar shock simulation.