Implemented a new momentum boundary class
BoundaryMomentumInjectRestrictSlab was ported from the main SPECTRUM repository. This class checks that a trajectory is within two parallel planes specified by the user before triggering a momentum boundary crossing. This new class is uploaded a test and should not change the results for the planar shock. If successful, a similar class will be created to implement in the runs for the spherical shock problem.