Update to input parameters
The list of input parameters for the simulations has been updated. Description of the new or changed parameters can be found in the latest version of README.md. In addition, a new momentum boundary class BoundaryMomentumInjectRestrictShell was ported from the original fork-parent SPECTRUM repository to handle momentum threshold crossings in the backward runs for the spherical case. A bug in DiffusionKineticEnergyRadialDistancePowerLaw was fixed as well.