Polished configuration/compilation scripts
The compilation and configuration scripts were polished for easier use, and the README was updated accordingly. A few minor edits were made in files, mostly regarding comments and indentation. One exception to this was the code dsa_backward_no_source.cc was edited to use a (zero) source term Parker transport module. The Makefile.am corresponding to this code was updated as well.